Check for libcrypt instead of unconditionally adding -lcrypt
This was causing the configure script to fail on checking for ncurses
when it was actually failing due to libcrypt being missing.  A check
for -lutil should be added to, but I didn't do this since -lutil is
not required since glibc 2.34 and I'm using a newer version and don't
know what function it should check for.